moved to the peg package rather than the tree package
